Changing weight data type to decimal, updating the MonthlyStatistics database query and setting dynamic min and max values of the LineChart

+++ Weight Tracker/ViewModels/BmiViewModel.cs
@@ -12,7 +12,7 @@ namespace Weight_Tracker.ViewModels
     {
         private readonly DateTime _today = Date.Today;
         private readonly DateTime _monthAgo = Date.MonthAgo;
-        public double BmiValue { get; set; }       
+        public decimal BmiValue { get; set; }       
 
         public BmiViewModel()
         {
